requirements.txt
autogen>=0.2.0
flask>=2.0.1
python-docx>=0.8.11
python-dotenv>=0.19.0
requests>=2.26.0
pillow>=8.3.1
matplotlib>=3.4.3
networkx>=2.6.3
pydantic>=1.8.2
python-multipart>=0.0.5
jinja2>=3.0.1
aiohttp>=3.8.1
openai>=1.0.0
graphviz>=0.17 